display tags interspersed chronologically with messages in message preview accessible from message review
