+++ src/components/IncomingMessageList/ConversationPreviewModal/index.jsx
@@ -1,120 +1,13 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import gql from 'graphql-tag'
-import { StyleSheet, css } from 'aphrodite'
-import moment from 'moment'
 import Dialog from 'material-ui/Dialog'
 import FlatButton from 'material-ui/FlatButton'
 
-import loadData from '../../containers//hoc/load-data'
-import wrapMutations from '../../containers/hoc/wrap-mutations'
-import MessageResponse from './MessageResponse';
-import ConversationLinkDialog from '../ConversationLinkDialog'
-import theme from '../../styles/theme'
-
-const styles = StyleSheet.create({
-  conversationRow: {
-    color: 'white',
-    padding: '10px',
-    borderRadius: '5px',
-    fontWeight: 'normal'
-  },
-  fromContact: {
-    marginLeft: undefined,
-    marginRight: '60px',
-    backgroundColor: '#AAAAAA'
-  },
-  fromTexter: {
-    marginLeft: '60px',
-    marginRight: undefined,
-    backgroundColor: 'rgb(33, 150, 243)'
-  },
-  tag: {
-    marginLeft: undefined,
-    marginRight: undefined,
-    backgroundColor: 'pink'
-  },
-  when: {
-    fontSize: theme.text.body.fontSize - 2
-  }
-})
-
-class MessageList extends Component {
-  componentDidMount() {
-    this.refs.messageWindow.scrollTo(0, this.refs.messageWindow.scrollHeight)
-  }
-
-  componentDidUpdate() {
-    this.refs.messageWindow.scrollTo(0, this.refs.messageWindow.scrollHeight)
-  }
-
-  render() {
-    const { messages, tags } = this.props
-    const items = [...messages, ...tags]
-    const sortedItems = items.sort((left, right) => right - left)
-    return  (
-      <div ref="messageWindow" style={{maxHeight: '300px', overflowY: 'scroll'}}>
-        {sortedItems.map((item, index) => {
-          let itemStyle = null
-          if (!!item.tag) {
-            itemStyle = styles.tag
-          } else {
-            const isFromContact = item.isFromContact
-            itemStyle = isFromContact ? styles.fromContact : styles.fromTexter
-          }
-
-          return (
-            <p key={index} className={css(styles.conversationRow, itemStyle)}>
-              {item.text || item.tag}
-              <br />
-              <span className={css(styles.when)}>{moment(item.createdAt).fromNow()}</span>
-            </p>
-          )
-        })}
-      </div>
-    )
-  }
-}
-
-MessageList.propTypes = {
-  messages: PropTypes.arrayOf(PropTypes.object),
-  tags: PropTypes.arrayOf(PropTypes.object)
-}
-
-class ConversationPreviewBody extends Component {
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      messages: props.conversation.messages,
-    }
-
-    this.messagesChanged = this.messagesChanged.bind(this)
-  }
-
-  messagesChanged(messages) {
-    this.setState({ messages })
-  }
-
-  render() {
-    return (
-      <div>
-        <MessageList
-          messages={this.state.messages}
-          tags={this.props.conversation.tags}
-        />
-        <MessageResponse
-          conversation={this.props.conversation}
-          messagesChanged={this.messagesChanged}
-        />
-      </div>
-    )
-  }
-}
-
-ConversationPreviewBody.propTypes = {
-  conversation: PropTypes.object
-}
+import loadData from '../../../containers/hoc/load-data'
+import wrapMutations from '../../../containers/hoc/wrap-mutations'
+import ConversationLinkDialog from '../../ConversationLinkDialog'
+import ConversationPreviewBody from './Body'
 
 class ConversationPreviewModal extends Component {
   constructor(props) {
@@ -157,8 +50,8 @@ class ConversationPreviewModal extends Component {
   }
 
   render() {
-    const { conversation } = this.props,
-          isOpen = conversation !== undefined
+    const { conversation } = this.props
+    const isOpen = conversation !== undefined
 
     const primaryActions = [
       <FlatButton
